requirements.txt
asgiref==3.11.1
asn1crypto==1.5.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
coverage==7.13.4
cryptography==46.0.5
cssselect2==0.9.0
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.1
idna==3.11
iniconfig==2.3.0
lxml==6.0.2
oscrypto==1.3.0
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
pycparser==3.0
pydyf==0.12.1
Pygments==2.19.2
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
PyJWT==2.11.0
pyphen==0.17.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-decouple==3.8
python-pkcs11==0.9.3
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
